Extract engine probe helpers behind the compatibility facade
After engine runtime helpers moved out, probe and get_duration were the smallest low-risk operation slice still embedded in engine.py. This moves them into engine_probe.py while re-exporting the exact function objects from mcp_video.engine for compatibility.

Constraint: Existing tests and callers import probe and get_duration from mcp_video.engine.

Rejected: Move normalize with probe | normalize performs transcoding and belongs in a separate operation-family slice.

Confidence: high

Scope-risk: narrow

Directive: Continue engine decomposition with small operation families while preserving engine.py as the public compatibility facade.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_probe.py

Tested: /opt/homebrew/bin/python3 probe/get_duration re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
